api/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
langchain==0.0.267
openai==0.28.1
python-dotenv==1.0.0
werkzeug==2.0.1
pypdf==3.15.4
faiss-cpu==1.7.4
tiktoken==0.5.1
chromadb==0.4.15
sentence-transformers>=2.2.2
